% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_get_data_function.R
\name{create_get_data_function}
\alias{create_get_data_function}
\title{Create a function to get a specific data set}
\usage{
create_get_data_function(id)
}
\arguments{
\item{id}{The ID of a data set}
}
\value{
A function that takes values of each dimension of the specified data set used to filter the data. These default to NULL, which should return the whole data set.
}
\description{
Returns a function that takes dimensions values as arguments which can be used to get and filter a specified data set
}
