% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{filter_dataset}
\alias{filter_dataset}
\title{Get a filtered data set}
\usage{
filter_dataset(id, ...)
}
\arguments{
\item{id}{ID of a data set. Data set IDs can be viewed using the `list_datasets` function.}

\item{...}{Further arguments specifying dimension values. Dimensions and possible values for a particular data set can be viewed using the `get_codelists` function.}
}
\value{
A data frame containing requested data.
}
\description{
Returns a data frame containing a filtered data set.
}
