% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_lists.R
\name{get_code_lists}
\alias{get_code_lists}
\title{Get code ONS code lists for specified data sets}
\usage{
get_code_lists(id)
}
\arguments{
\item{id}{A data set id}
}
\value{
List of code lists relevant to the specified data set
}
\description{
Get code ONS code lists for specified data sets
}
