% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Get a data set}
\usage{
get_dataset(id, full_dataset = FALSE, ...)
}
\arguments{
\item{id}{ID of a data set. Data set IDs can be viewed using the `list_datasets` function.}

\item{full_dataset}{A logical value indicating whether you want to download the full data set. If `TRUE` any further parameters are ignored.}

\item{...}{Further arguments specifying dimension values. Dimensions and possible values for a particular data set can be viewed using the `get_codelists` function.}
}
\value{
A data frame containing requested data.
}
\description{
Returns a data frame containing a data set.
}
