% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{List data sets available from the ONS}
\usage{
list_datasets(include_desc = FALSE)
}
\arguments{
\item{include_desc}{A logical value indicating whether the descriptions
of each dataset should be included.}
}
\value{
A data frame of data set IDs.
}
\description{
Returns a data frame listing the ids of data sets available from the ONS
}
